requirements.txt
blinker==1.4
certifi==2018.11.29
chardet==3.0.4
Cython==0.28.2
docutils==0.14
idna==2.7
Kivy==1.10.1
Kivy-Garden==0.1.4
mongoengine==0.15.3
numpy==1.15.4
Pygments==2.3.0
pymongo==3.7.1
requests==2.20.1
six==1.11.0
urllib3==1.24.1
websocket-client==0.53.0
